worker/requirements.txt
supabase
anthropic
google-api-python-client
requests
python-dotenv
